Add a config option sub_genres
If this is set to False, the playlists for a genre will be returned for
the genre uri, instead of a link to that, a link to the top tracks
playlist and the sub genres. This makes the number of levels to navigate
down to the genre playlists one shorter.

This change is motived by the fact that the API doesn't seem to return
any top tracks playlists or more than one sub genre, so those entries
doesn't have much value anymore. Nevertheless, the default is still to
show them, so the default behavior doesn't change.